Fix unused file inclusion and deprecated API usage
Removed included "Vec.hpp" file (see https://github.com/JayhawkZombie/RayCasting/issues/1), as the class was unused in code.

Replaced use of deprecated SFML API's 'setColor' with 'setFillColor' for drawn primitives, and 'setParameter' with 'setUniform' for shader parameters.

Added include for '<memory>' header, since 'LightSystem.h' did not include it directly.

Added definition for 'FLOAT_MAX' since it was referenced in 'LightSystem.h' and originally defined by the now-unused Vec.hpp header.
This is, of course, not a real max, but a max set high enough to eliminate matches in the algorithm since a distance of '1E10' is far outside the distance considered to be close enough for lights to hit.

NOTE: Consider not using a lookup table as is demonstrated in the LightSystem header.  After testing, it did not provide any noticeable benefit for marginally small table sizes and presented a performance degradation for larger tables.  That is likely from c